test(sagemaker): consolidate tag tests and fix linting
- Merge sagemaker_service_tags_test.py into sagemaker_service_test.py as requested by reviewer.
- Fix flake8 linting errors (unused imports and shadowing).
- Apply black formatting to test files.
- Add entry to prowler/CHANGELOG.md for parallel tag retrieval.
